requirements.txt
python-dotenv
python-fasthtml
monsterUi
pydantic
email-validator